// src/app/shared/app/models/Production/i-motor-active-preview.ts
export interface IMotorActiveDataPreview {
  sNo: number;
  oasisPolRef: string;
  policiesSNo: number;
  policyNo: string;
  clientID: number;
  valid: number;
  moveNo: number;
  ownerDriver: string;
  plateNo: string;
  plateChar1: string;
  plateChar2: string;
  plateChar3: string;
  sequenceNo: string;
  customID: string;
  brandName: string;
  model: string;
  inception: Date;
  expiry: Date;
  kclDate: Date;
  seats: string;
  bodyType: string;
  chassisNo: string;
  marketValue: string;
  repairType: string;
  vehicleOwnerID: string;
  projectName: string;
  savedUser: string;
  savedOn: Date;
}